slack-bot-demo/requirements.txt
slack-bolt>=1.18
requests>=2.31
python-dotenv>=1.0
boto3>=1.34
